requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
aws-requests-auth==0.4.0
beautifulsoup4==4.8.2
black==18.9b0
boto3==1.9.189
botocore==1.12.33
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cache-url==2.0.0
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-filter==1.1.0
django-modelcluster==5.0.1
django-modeltranslation==0.15
django-redis==4.11.0
django-storages==1.8
django-taggit==1.2.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.2
djangorestframework==3.11.0
docutils==0.16
draftjs-exporter==2.1.7
draftjs-exporter-markdown==0.2.1
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
jedi==0.15.2
jmespath==0.9.5
l18n==2018.5
Markdown==3.2.2
mypy==0.770
mypy-extensions==0.4.3
oauthlib==3.1.0
parso==0.7.0
pathspec==0.8.0
Pillow==6.2.2
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.4
pudb==2019.2
Pygments==2.6.1
pyls-mypy==0.1.8
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python3-openid==3.1.0
pytz==2020.1
redis==3.4.1
regex==2020.5.13
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
ujson==1.35
Unidecode==1.1.1
urllib3==1.24.3
urwid==2.1.0
wagtail==2.9
wagtailfontawesome==1.1.4
wagtailmenus==2.8
wagtailtrans==2.1
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.8
zipp==3.1.0
